Move single-use functions to calling classes
