+++ app/Http/Controllers/ViewTestController.php
@@ -4,6 +4,7 @@ namespace App\Http\Controllers;
 
 use CDash\Controller\Api\ViewTest;
 use CDash\Database;
+use CDash\Model\Build;
 use Illuminate\Http\JsonResponse;
 use Illuminate\View\View;
 use Symfony\Component\HttpFoundation\StreamedResponse;
@@ -50,7 +51,7 @@ final class ViewTestController extends AbstractBuildController
     public function fetchPageContent(): JsonResponse|StreamedResponse
     {
         $db = Database::getInstance();
-        $controller = new ViewTest($db, get_request_build());
+        $controller = new ViewTest($db, self::get_request_build());
         $response = $controller->getResponse();
         if ($controller->JSONEncodeResponse) {
             return response()->json(cast_data_for_JSON($response));
@@ -62,5 +63,55 @@ final class ViewTestController extends AbstractBuildController
                 echo $response;
             }, 'test-export.csv', $headers);
         }
+    }
+
+    /**
+     * Returns a build based on the id extracted from the request and returns it if the user has
+     * necessary access to the project
+     *
+     * @param bool $required
+     */
+    private static function get_request_build($required = true): ?Build
+    {
+        $id = self::get_request_build_id($required);
+        if (is_null($id)) {
+            return null;
+        }
+        $build = new Build();
+        $build->Id = $id;
+
+        if ($required && !$build->Exists()) {
+            abort(400, 'This build does not exist. Maybe it has been deleted.');
+        }
+
+        if ($id) {
+            $build->FillFromId($id);
+        }
+
+        return can_access_project($build->ProjectId) ? $build : null;
+    }
+
+    /**
+     * Pulls the buildid from the request
+     *
+     * @param bool $required
+     */
+    private static function get_request_build_id($required = true): ?int
+    {
+        $buildid = self::get_int_param('buildid', $required);
+        return $buildid;
+    }
+
+    private static function get_int_param($name, $required = true): ?int
+    {
+        $value = get_param($name, $required);
+        if (is_null($value)) {
+            return null;
+        }
+
+        if ($required && !is_numeric($value)) {
+            abort(400, "Valid $name required");
+        }
+        return (int) $value;
     }
 }
