+++ phpstan-baseline.neon
@@ -2710,6 +2710,30 @@ parameters:
 			path: app/Http/Controllers/UserNoteController.php
 
 		-
+			message: '#^Call to function is_null\(\) with string will always evaluate to false\.$#'
+			identifier: function.impossibleType
+			count: 1
+			path: app/Http/Controllers/ViewTestController.php
+
+		-
+			message: '#^Method App\\Http\\Controllers\\ViewTestController\:\:get_int_param\(\) has parameter \$name with no type specified\.$#'
+			identifier: missingType.parameter
+			count: 1
+			path: app/Http/Controllers/ViewTestController.php
+
+		-
+			message: '#^Method App\\Http\\Controllers\\ViewTestController\:\:get_int_param\(\) has parameter \$required with no type specified\.$#'
+			identifier: missingType.parameter
+			count: 1
+			path: app/Http/Controllers/ViewTestController.php
+
+		-
+			message: '#^Only booleans are allowed in an if condition, int given\.$#'
+			identifier: if.condNotBoolean
+			count: 1
+			path: app/Http/Controllers/ViewTestController.php
+
+		-
 			message: '#^Parameter \#2 \$build of class CDash\\Controller\\Api\\ViewTest constructor expects CDash\\Model\\Build, CDash\\Model\\Build\|null given\.$#'
 			identifier: argument.type
 			count: 1
@@ -14842,12 +14866,6 @@ parameters:
 			path: app/cdash/include/api_common.php
 
 		-
-			message: '#^Call to function is_null\(\) with string will always evaluate to false\.$#'
-			identifier: function.impossibleType
-			count: 1
-			path: app/cdash/include/api_common.php
-
-		-
 			message: '#^Construct empty\(\) is not allowed\. Use more strict comparison\.$#'
 			identifier: empty.notAllowed
 			count: 2
@@ -14884,44 +14902,14 @@ parameters:
 			path: app/cdash/include/api_common.php
 
 		-
-			message: '#^Function get_int_param\(\) has parameter \$name with no type specified\.$#'
-			identifier: missingType.parameter
-			count: 1
-			path: app/cdash/include/api_common.php
-
-		-
-			message: '#^Function get_int_param\(\) has parameter \$required with no type specified\.$#'
-			identifier: missingType.parameter
-			count: 1
-			path: app/cdash/include/api_common.php
-
-		-
 			message: '#^Function get_param\(\) has parameter \$name with no type specified\.$#'
 			identifier: missingType.parameter
 			count: 1
 			path: app/cdash/include/api_common.php
 
 		-
-			message: '#^Left side of && is always true\.$#'
-			identifier: booleanAnd.leftAlwaysTrue
-			count: 1
-			path: app/cdash/include/api_common.php
-
-		-
-			message: '#^Only booleans are allowed in &&, CDash\\Model\\Project given on the left side\.$#'
-			identifier: booleanAnd.leftNotBoolean
-			count: 1
-			path: app/cdash/include/api_common.php
-
-		-
 			message: '#^Only booleans are allowed in a negated boolean, mixed given\.$#'
 			identifier: booleanNot.exprNotBoolean
-			count: 1
-			path: app/cdash/include/api_common.php
-
-		-
-			message: '#^Only booleans are allowed in an if condition, int given\.$#'
-			identifier: if.condNotBoolean
 			count: 1
 			path: app/cdash/include/api_common.php
 
@@ -16366,9 +16354,21 @@ parameters:
 			path: app/cdash/public/api/v1/buildgroup.php
 
 		-
+			message: '#^Left side of && is always true\.$#'
+			identifier: booleanAnd.leftAlwaysTrue
+			count: 1
+			path: app/cdash/public/api/v1/buildgroup.php
+
+		-
 			message: '#^Loose comparison via "\=\=" is not allowed\.$#'
 			identifier: equal.notAllowed
 			count: 3
+			path: app/cdash/public/api/v1/buildgroup.php
+
+		-
+			message: '#^Only booleans are allowed in &&, CDash\\Model\\Project given on the left side\.$#'
+			identifier: booleanAnd.leftNotBoolean
+			count: 1
 			path: app/cdash/public/api/v1/buildgroup.php
 
 		-
